moved waitForEvent() and wait for load before entering webXR
