+++ packages/model-viewer/src/test/features/staging-spec.ts
@@ -17,7 +17,8 @@ import {CameraChangeDetails} from '../../features/controls.js';
 import {StagingMixin} from '../../features/staging.js';
 import ModelViewerElementBase from '../../model-viewer-base.js';
 import {ChangeSource} from '../../three-components/SmoothControls.js';
-import {assetPath, rafPasses, timePasses, waitForEvent} from '../helpers.js';
+import {waitForEvent} from '../../utilities.js';
+import {assetPath, rafPasses, timePasses,} from '../helpers.js';
 import {BasicSpecTemplate} from '../templates.js';
 
 const expect = chai.expect;
