+++ packages/model-viewer/src/test/helpers.ts
@@ -12,11 +12,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {EventDispatcher, Group, Texture} from 'three';
+import {Group, Texture} from 'three';
 import {GLTFParser} from 'three/examples/jsm/loaders/GLTFLoader.js';
 
 import {ExpressionNode, ExpressionTerm, FunctionNode, HexNode, IdentNode, Operator, OperatorNode} from '../styles/parsers.js';
-import {deserializeUrl} from '../utilities.js';
+import {deserializeUrl, PredicateFunction} from '../utilities.js';
 
 export const elementFromLocalPoint =
     (document: Document|ShadowRoot, x: number, y: number): Element|null => {
@@ -40,14 +40,6 @@ export const pickShadowDescendant =
 
 export const timePasses = (ms: number = 0): Promise<void> =>
     new Promise(resolve => setTimeout(resolve, ms));
-
-export type PredicateFunction<T = void> = (value: T) => boolean;
-
-/**
- * Three.js EventDispatcher and DOM EventTarget use different event patterns,
- * so AnyEvent covers the shape of both event types.
- */
-export type AnyEvent = Event|CustomEvent<any>|{[index: string]: string};
 
 export const until =
     async (predicate: PredicateFunction) => {
@@ -74,25 +66,6 @@ export const textureMatchesMeta =
            Object.keys(meta).reduce((matches, key) => {
              return matches && meta[key] === (texture as any).userData[key];
            }, true));
-
-/**
- * @param {EventTarget|EventDispatcher} target
- * @param {string} eventName
- * @param {?Function} predicate
- */
-export const waitForEvent = <T extends AnyEvent = Event>(
-    target: EventTarget|EventDispatcher,
-    eventName: string,
-    predicate: PredicateFunction<T>|null = null): Promise<T> =>
-    new Promise(resolve => {
-      function handler(event: AnyEvent) {
-        if (!predicate || predicate(event as T)) {
-          resolve(event as T);
-          target.removeEventListener(eventName, handler);
-        }
-      }
-      target.addEventListener(eventName, handler);
-    });
 
 export interface SyntheticEventProperties {
   clientX?: number;
