+++ packages/model-viewer/src/test/utilities/focus-visible-spec.ts
@@ -14,8 +14,9 @@
  */
 
 import ModelViewerElementBase from '../../model-viewer-base.js';
+import {waitForEvent} from '../../utilities.js';
 import {FocusVisiblePolyfillMixin} from '../../utilities/focus-visible.js';
-import {assetPath, waitForEvent} from '../helpers.js';
+import {assetPath} from '../helpers.js';
 import {BasicSpecTemplate} from '../templates.js';
 
 const expect = chai.expect;
