+++ packages/model-viewer/src/utilities.ts
@@ -13,6 +13,7 @@
  * limitations under the License.
  */
 
+import {EventDispatcher} from 'three';
 import {HAS_WEBXR_DEVICE_API, HAS_WEBXR_HIT_TEST_API, IS_WEBXR_AR_CANDIDATE} from './constants.js';
 
 export type Constructor<T = object, U = object> = {
@@ -225,3 +226,30 @@ export const getFirstMapKey = <T = any, U = any>(map: Map<T, U>): T|null => {
 
   return firstKey;
 };
+
+/**
+ * Three.js EventDispatcher and DOM EventTarget use different event patterns,
+ * so AnyEvent covers the shape of both event types.
+ */
+export type AnyEvent = Event|CustomEvent<any>|{[index: string]: string};
+
+export type PredicateFunction<T = void> = (value: T) => boolean;
+
+/**
+ * @param {EventTarget|EventDispatcher} target
+ * @param {string} eventName
+ * @param {?Function} predicate
+ */
+export const waitForEvent = <T extends AnyEvent = Event>(
+    target: EventTarget|EventDispatcher,
+    eventName: string,
+    predicate: PredicateFunction<T>|null = null): Promise<T> =>
+    new Promise(resolve => {
+      function handler(event: AnyEvent) {
+        if (!predicate || predicate(event as T)) {
+          resolve(event as T);
+          target.removeEventListener(eventName, handler);
+        }
+      }
+      target.addEventListener(eventName, handler);
+    });
