+++ packages/modelviewer.dev/examples/loading/index.html
@@ -315,7 +315,7 @@ ModelViewerElement.dracoDecoderLocation = 'http://example.com/location/of/draco/
           </div>
           <example-snippet stamp-to="demo-container-9" highlight-as="html">
             <template>
-<model-viewer camera-controls alt="A 3D model of an astronaut" ios-src="../../shared-assets/models/Astronaut.usdz">
+<model-viewer camera-controls alt="A 3D model of an astronaut" ar ios-src="../../shared-assets/models/Astronaut.usdz">
 </model-viewer>
             </template>
           </example-snippet>
