+++ packages/modelviewer.dev/examples/model-formats.html
@@ -167,7 +167,7 @@ ModelViewerElement.dracoDecoderLocation =
         </div>
         <example-snippet stamp-to="demo-container-4" highlight-as="html">
           <template>
-<model-viewer camera-controls alt="A 3D model of an astronaut" ios-src="../shared-assets/models/Astronaut.usdz">
+<model-viewer camera-controls alt="A 3D model of an astronaut" ar ios-src="../shared-assets/models/Astronaut.usdz">
 </model-viewer>
           </template>
         </example-snippet>
